generalize module included in all delayed jobs
now also applies to other activities one might want to execute before all delayed jobs
(as discussed with @ulferts)
