+++ app/workers/copy_project_job.rb
@@ -34,7 +34,7 @@ class CopyProjectJob < Struct.new(:user_id,
                                   :associations_to_copy,
                                   :send_mails)
   include OpenProject::LocaleHelper
-  include OpenProject::ResetsRequestStore
+  include OpenProject::BeforeDelayedJob
 
   def perform
     User.current = user
