+++ app/workers/delete_user_job.rb
@@ -28,7 +28,7 @@
 #++
 
 class DeleteUserJob
-  include OpenProject::ResetsRequestStore
+  include OpenProject::BeforeDelayedJob
 
   def initialize(user)
     @user_id = user.id
