+++ app/workers/deliver_work_package_notification_job.rb
@@ -28,7 +28,7 @@
 #++
 
 class DeliverWorkPackageNotificationJob
-  include OpenProject::ResetsRequestStore
+  include OpenProject::BeforeDelayedJob
 
   def initialize(journal_id, author_id)
     @journal_id = journal_id
