+++ app/workers/enqueue_work_package_notification_job.rb
@@ -29,7 +29,7 @@
 
 # Enqueues
 class EnqueueWorkPackageNotificationJob
-  include OpenProject::ResetsRequestStore
+  include OpenProject::BeforeDelayedJob
 
   def initialize(journal_id, author_id)
     @journal_id = journal_id
