+++ app/workers/scm/delete_repository_job.rb
@@ -35,7 +35,7 @@
 # creation and deletion of repositories BOTH on the database and filesystem.
 # Until then, a synchronous process is more failsafe.
 class Scm::DeleteRepositoryJob
-  include OpenProject::ResetsRequestStore
+  include OpenProject::BeforeDelayedJob
 
   def initialize(managed_path)
     @managed_path = managed_path
